+++ src/components/molecules/ExecutionDetails/ExecutionDetails.styled.ts
@@ -112,29 +112,3 @@ export const StyledTestSuiteExecutionResultsListItem = styled.li`
 
   color: ${Colors.whitePure};
 `;
-
-export const ItemRow = styled.div<{$flex: number}>`
-  display: flex;
-  align-items: center;
-  flex: ${({$flex}) => $flex};
-  justify-content: space-between;
-
-  &:not(:last-child) {
-    margin-bottom: 10px;
-  }
-`;
-
-export const ItemColumn = styled.div`
-  overflow: hidden;
-
-  display: flex;
-  flex-wrap: wrap;
-  align-items: center;
-  gap: 16px;
-
-  flex: 1;
-
-  &.flex-auto {
-    flex: 0 1 auto;
-  }
-`;
