+++ src/components/molecules/ExecutionDetails/index.ts
@@ -1,1 +1,2 @@
-export {default} from './ExecutionDetails';
+export {default as TestExecutionDetailsTabs} from './TestExecutionDetails/TestExecutionDetailsTabs';
+export {default as TestSuiteExecutionDetailsTabs} from './TestSuiteExecutionDetails/TestSuiteExecutionDetailsTabs';
