+++ src/components/molecules/index.ts
@@ -13,7 +13,7 @@ export {default as EntityGrid} from './EntityGrid';
 export {default as notificationCall} from './Notification';
 export {default as EndpointModal} from './EndpointModal';
 export {default as ConfigurationCard} from './ConfigurationCard';
-export {default as ExecutionDetails} from './ExecutionDetails';
+export {TestExecutionDetailsTabs, TestSuiteExecutionDetailsTabs} from './ExecutionDetails';
 export {default as DragNDropList} from './DragNDropList';
 export {default as ArtifactsList} from './ArtifactsList';
 export {default as TestSuiteStepCard} from './TestSuiteStepCard';
