+++ src/components/organisms/EntityDetails/ExecutionDetailsDrawer/ExecutionDetailsDrawer.styled.tsx
@@ -23,27 +23,55 @@ export const ExecutionDetailsDrawerWrapper = styled(motion.div)<{$isRowSelected:
   // flex: ${props => getInfoPanelFlexProperty(props)};
 
   height: 100%;
-  width: ${props => (props.drawerWidth ? `${props.drawerWidth}` : '85vw')}
+  width: ${props => (props.drawerWidth ? `${props.drawerWidth}` : '85vw')};
 
   margin: 0;
   padding: ${({$isRowSelected}) => ($isRowSelected ? '45px 30px' : '0')};
 
   background: ${Colors.slate800};
-
-  transition: 0.3s flex;
 `;
 
 export const DrawerHeader = styled.div`
   display: flex;
-
-  padding-bottom: 50px;
 `;
+
+export const DrawerContent = styled.div``;
 
 export const ItemRow = styled.div<{$flex: number}>`
   display: flex;
   align-items: center;
   flex: ${({$flex}) => $flex};
   justify-content: space-between;
+
+  &:not(:last-child) {
+    margin-bottom: 10px;
+  }
 `;
 
-export const DrawerContent = styled.div``;
+export const ItemColumn = styled.div`
+  overflow: hidden;
+
+  display: flex;
+  flex-wrap: wrap;
+  align-items: center;
+  gap: 16px;
+
+  flex: 1;
+
+  &.flex-auto {
+    flex: 0 1 auto;
+  }
+`;
+
+export const HeaderContent = styled.div`
+  overflow: hidden;
+
+  display: flex;
+  flex-direction: column;
+  flex: 1;
+  gap: 12px;
+
+  margin-left: 15px;
+
+  width: 0;
+`;
